+++ org.gitools.heatmap/src/main/java/org/gitools/heatmap/header/HierarchicalClusterHeatmapHeader.java
@@ -23,6 +23,7 @@ package org.gitools.heatmap.header;
 
 import com.google.common.base.Function;
 import com.google.common.collect.Lists;
+import org.apache.commons.math3.util.FastMath;
 import org.gitools.heatmap.HeatmapDimension;
 import org.gitools.heatmap.decorator.DetailsDecoration;
 import org.gitools.matrix.model.matrix.AnnotationMatrix;
@@ -31,7 +32,11 @@ import org.gitools.utils.textpattern.TextPattern;
 import javax.xml.bind.annotation.*;
 import java.awt.*;
 import java.util.ArrayList;
+import java.util.HashMap;
 import java.util.List;
+import java.util.Map;
+import java.util.regex.Matcher;
+import java.util.regex.Pattern;
 
 @XmlAccessorType(XmlAccessType.NONE)
 @XmlRootElement
@@ -48,6 +53,21 @@ public class HierarchicalClusterHeatmapHeader extends HeatmapHeader {
     @XmlTransient
     private boolean reportLastInteraction;
 
+    public static final String PROPERTY_VISIBLE_LEVELS = "visibleLevels";
+    @XmlTransient
+    private String visibleLevels;
+
+    @XmlTransient
+    private String visibleLevelsFormat = "(\\d+(\\-\\d+)?(,(?!$))?)+";
+
+    public static final String PROPERTY_MAX_LEVELS = "maxLevels";
+    @XmlTransient
+    private int maxLevels = -1;
+
+    public static final String PROPERTY_COLOR_PALETTE = "colorPalette";
+    @XmlElement
+    private String colorPalette = "Default";
+
     @XmlElement
     private HierarchicalCluster hierarchicalCluster;
 
@@ -60,9 +80,62 @@ public class HierarchicalClusterHeatmapHeader extends HeatmapHeader {
         super(hdim);
         clusterLevels = new ArrayList<>();
         reportLastInteraction = false;
+        setMargin(1);
+    }
+
+    public static void createHierarchicalLevelsHeaders(HierarchicalClusterHeatmapHeader hierarchicalHeader,
+                                                       int maxLevels,
+                                                       String annotationPrefix) {
+        int currentLevel = 0;
+        HeatmapDimension clusteringDimension = hierarchicalHeader.getHeatmapDimension();
+        AnnotationMatrix annotationMatrix = clusteringDimension.getAnnotations();
+        Map<Integer, List<HierarchicalCluster>> clustersMapPerLevel = new HashMap<>();
+        List<HierarchicalCluster> children = hierarchicalHeader.getHierarchicalCluster().getChildren();
+        while (currentLevel < maxLevels) {
+            currentLevel++;
+
+            List<HierarchicalCluster> nextLevel = new ArrayList<>();
+            for (HierarchicalCluster cluster : children) {
+                if (!cluster.getChildren().isEmpty()) {
+                    for (String identifier : cluster.getIdentifiers()) {
+                        annotationMatrix.setAnnotation(identifier, annotationPrefix + currentLevel, cluster.getName());
+                    }
+                }
+                nextLevel.addAll(cluster.getChildren());
+            }
+
+            clustersMapPerLevel.put(currentLevel, children);
+            children = nextLevel;
+
+            if (children.isEmpty()) {
+                currentLevel--;
+                break;
+            }
+        }
+
+        // Hierarchical clustering headers
+        int depth = FastMath.min(maxLevels, currentLevel);
+        for (int l = depth; l >= 1; l--) {
+            HeatmapColoredLabelsHeader levelHeader = new HeatmapColoredLabelsHeader(clusteringDimension);
+
+            List<HierarchicalCluster> clusters = clustersMapPerLevel.get(l);
+            List<ColoredLabel> coloredLabels = new ArrayList<>(clusters.size());
+            for (HierarchicalCluster cluster : clusters) {
+                coloredLabels.add(new ColoredLabel(cluster.getName(), new Color(cluster.getColor())));
+            }
+
+            levelHeader.setClusters(coloredLabels);
+
+
+            levelHeader.setTitle(annotationPrefix + l);
+            levelHeader.setSize(7);
+            levelHeader.setAnnotationPattern("${" + annotationPrefix + l + "}");
+            hierarchicalHeader.addLevel(levelHeader);
+        }
     }
 
     public void addLevel(HeatmapColoredLabelsHeader level) {
+        level.setMargin(1);
         clusterLevels.add(level);
     }
 
@@ -72,6 +145,14 @@ public class HierarchicalClusterHeatmapHeader extends HeatmapHeader {
         for (HeatmapColoredLabelsHeader levels : clusterLevels) {
             levels.init(heatmapDimension);
         }
+    }
+
+    @Override
+    public void setMargin(int margin) {
+        for (HeatmapColoredLabelsHeader levelHeader : getClusterLevels()) {
+            levelHeader.setMargin(margin);
+        }
+        super.setMargin(margin);
     }
 
     public List<HeatmapColoredLabelsHeader> getClusterLevels() {
@@ -130,7 +211,9 @@ public class HierarchicalClusterHeatmapHeader extends HeatmapHeader {
     public int getSize() {
         size = 0;
         for (HeatmapColoredLabelsHeader level : clusterLevels) {
-            size += level.getSize();
+            if (level.isVisible()) {
+                size += level.getSize();
+            }
         }
         return size;
     }
@@ -222,7 +305,7 @@ public class HierarchicalClusterHeatmapHeader extends HeatmapHeader {
      */
     public void setSize(int size) {
 
-        int change =0;
+        int change = 0;
         if (this.size < size) {
             change = 1;
         } else if (this.size > size) {
@@ -245,6 +328,130 @@ public class HierarchicalClusterHeatmapHeader extends HeatmapHeader {
 
     }
 
+    public String getVisibleLevels() {
+
+        if (visibleLevels == null) {
+            visibleLevels = generateVisibleLevelsString();
+        }
+        return visibleLevels;
+    }
+
+    public void setVisibleLevels(String visibleLevels) {
+
+        Pattern p = Pattern.compile(visibleLevelsFormat);
+        Matcher m = p.matcher(visibleLevels);
+        if (m.matches()) {
+            //System.out.println("Valid pattern: " + visibleLevels);
+
+            String old = this.visibleLevels;
+            this.visibleLevels = visibleLevels;
+
+            setLevelHeadersVisibility(visibleLevels);
+            firePropertyChange(PROPERTY_VISIBLE_LEVELS, old, visibleLevels);
+        }
+
+
+    }
+
+    public int getMaxLevels() {
+        if (maxLevels == -1) {
+            maxLevels = getClusterLevels().size();
+        }
+        return maxLevels;
+    }
+
+    public void setMaxLevels(int maxLevels) {
+        int old = this.maxLevels;
+        if (maxLevels < 10) {
+            maxLevels = 10;
+        }
+        if (maxLevels != this.maxLevels) {
+            String annoPrefix = getClusterLevels().get(1).getTitle().split(" L")[0];
+            removeMetadata();
+            HierarchicalClusterHeatmapHeader.createHierarchicalLevelsHeaders(this, maxLevels, annoPrefix);
+            this.maxLevels = maxLevels;
+            firePropertyChange(PROPERTY_MAX_LEVELS, old, maxLevels);
+
+        }
+    }
+
+    public String getColorPalette() {
+        if (colorPalette == null) {
+            colorPalette = "Default";
+        }
+        return colorPalette;
+    }
+
+    public void setColorPalette(String colorPalette) {
+        String old = this.colorPalette;
+        this.colorPalette = colorPalette;
+        if (!old.equals(colorPalette)) {
+            // repaint
+            String annoPrefix = getClusterLevels().get(1).getTitle().split("1")[0];
+            removeMetadata();
+            String oldname = this.hierarchicalCluster.getName();
+            hierarchicalCluster.setName("");
+            HierarchicalClusterNamer.nameClusters(this.hierarchicalCluster, colorPalette);
+            hierarchicalCluster.setName(oldname);
+            HierarchicalClusterHeatmapHeader.createHierarchicalLevelsHeaders(this, maxLevels, annoPrefix);
+            firePropertyChange(PROPERTY_COLOR_PALETTE, old, colorPalette);
+        }
+    }
+
+
+    private void setLevelHeadersVisibility(String visibleLevels) {
+
+        List<Integer> visibleList = new ArrayList<>();
+        for (String section : visibleLevels.split(",")) {
+            if (!section.contains("-")) {
+                visibleList.add(Integer.parseInt(section));
+            } else {
+                int start = Integer.parseInt(section.split("-")[0]);
+                int end = Integer.parseInt(section.split("-")[1]);
+                for (int i = start; i <= end; i++) {
+                    visibleList.add(i);
+                }
+            }
+        }
+
+        int level = 0;
+        for (HeatmapColoredLabelsHeader levelHeader : Lists.reverse(getClusterLevels())) {
+            level++;
+            boolean isVisible = visibleList.contains(level);
+            levelHeader.setVisible(isVisible);
+        }
+    }
+
+    private String generateVisibleLevelsString() {
+        int level = 1;
+        String visibleLevels = "";
+        int lastReportedLevel = -1;
+        int lastVisible = -1;
+        for (HeatmapColoredLabelsHeader labelsHeader : getClusterLevels()) {
+
+            if (labelsHeader.isVisible() && lastReportedLevel == -1) {
+                visibleLevels = Integer.toString(level);
+                lastReportedLevel = level;
+                lastVisible = level;
+            } else if (labelsHeader.isVisible()) {
+                lastVisible = level;
+            } else {
+                if (lastVisible == lastReportedLevel) {
+                    visibleLevels += ",";
+                } else {
+                    visibleLevels += "-" + Integer.toString(lastVisible) + ",";
+                }
+            }
+            level++;
+        }
+
+        // after last loop
+        if (lastVisible != lastReportedLevel) {
+            visibleLevels += "-" + Integer.toString(lastVisible);
+        }
+        return visibleLevels;
+    }
+
     public void removeMetadata() {
         AnnotationMatrix annotation = getHeatmapDimension().getAnnotations();
         for (HeatmapColoredLabelsHeader level : getClusterLevels()) {
@@ -259,4 +466,5 @@ public class HierarchicalClusterHeatmapHeader extends HeatmapHeader {
         clusterLevels.clear();
         annotation.getLabels();
     }
+
 }